requirements-gpu.txt
toml
h5py
scikit-learn
tensorflow-gpu
keras